Use consistent numeric types for sample-domain quantities
Report the field disk location in f32, and track the running input
offset as an integer sample index plus a sub-sample fractional
remainder. Type the absolute sample positions (read, scheduled, and
file locations, plus the block start) and the per-field sample span as
u64, since they are never negative.

+++ crates/tape-decode/src/decode/field.rs
@@ -731,14 +731,14 @@ pub(crate) fn predecode_field_from_rawdecode(
     spec: &DecoderSpec,
     scheduled_prevfield: Option<PrevFieldState>,
     inter_field_state: &mut InterFieldState,
-    scheduled_readloc: i64,
+    scheduled_readloc: u64,
     resync_state: &mut ResyncState,
     chroma_afc_state: &ChromaAfcState,
 ) -> Result<DecodeFieldResult> {
     // Build and classify the next DecodedField from coalesced block data. This is
     // the sync/line-location half of the speculative predecode step; output/TBC
     // conversion and metadata writing remain in the executable orchestrator.
-    let readloc = rawdecode.startloc as i64;
+    let readloc = rawdecode.startloc;
     let mut pending_field = DecodedField {
         data: rawdecode,
         prevfield: scheduled_prevfield,
